+++ noodles-bgzf/src/async.rs
@@ -5,17 +5,11 @@ pub mod io;
 
 use self::block_codec::BlockCodec;
 
-#[deprecated(since = "0.38.0", note = "Use `bgzf::r#async::io::Reader` instead.")]
-pub use self::io::Reader;
-
-#[deprecated(since = "0.38.0", note = "Use `bgzf::r#async::io::Writer` instead.")]
-pub use self::io::Writer;
-
 #[cfg(test)]
 mod tests {
     use tokio::io::{self, AsyncReadExt, AsyncWriteExt};
 
-    use super::*;
+    use super::io::{Reader, Writer};
 
     #[tokio::test]
     async fn test_self() -> io::Result<()> {