+++ noodles-bgzf/src/gzi.rs
@@ -12,13 +12,3 @@ mod index;
 pub mod io;
 
 pub use self::index::Index;
-
-#[deprecated(since = "0.35.0", note = "Use `gzi::fs::read` instead.")]
-pub use self::fs::read;
-
-#[deprecated(since = "0.35.0", note = "Use `gzi::io::Reader` instead.")]
-pub use self::io::Reader;
-
-#[cfg(feature = "async")]
-#[deprecated(since = "0.35.0", note = "Use `bgzf::gzi::r#async::Reader` instead.")]
-pub use self::r#async::Reader as AsyncReader;